requirements.txt
datasets==3.6.0
huggingface_hub==0.31.1
langchain_core==0.3.66
langchain_ollama==0.3.3
matplotlib==3.10.3
nltk==3.9.1
pandas==2.3.0
PyYAML==6.0.2
PyYAML==6.0.2
sentence_transformers==4.1.0
spacy==3.8.5
tabulate==0.9.0
torch==2.7.0+cu128
